requirements.txt
numpy
pandas
scikit-learn
paho-mqtt
fastapi
uvicorn
joblib